core: initial drop
tmux-set-buffer-remote is my own attempt to get rid of the system
clipboard on Linux.  To do so, I use tmux's paste buffer instead.
Syncing paste buffers from a remote machine might be tricky, which is
why this small tool was born.

To use it in tmux, you can add the following line in your .tmux.conf:
    bind-key -t vi-copy 'y' copy-pipe 'tmux-set-buffer-remote user@host'

This assumes that you use vi-copy mode.

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>